requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=1.51.0
google-cloud-storage>=2.18.0
google-cloud-texttospeech>=2.17.0
python-multipart>=0.0.12
pyyaml>=6.0
